chore: move tasks to core folder
